Fix offset issue when reading StringFileInfo
This changes the string parsing logic to add the length of the alignment
padding to the size of the string when reporting the size of the
StringFileInfo structure. Previously, the caller did not keep track of
this, and would end reading the next string a few bytes short.
This was generally fixed by the alignment code except for the case where
the offset was short by four bytes, leading to the code trying to read
the string from the wrong offset leading to corrupted data.