feat: parse a large number of .NET metadata headers (#86)
* feat: parse a large number of .NET metadata headers

Support parsing for most of the metadata headers described in ECMA 335 and update/add tests for them. This is intentionally structured to minimize drastic changes to the existing .NET parsing code.

* feat: parse a large number of .NET metadata headers

Support parsing for most of the metadata headers described in ECMA 335 and update/add tests for them. This is intentionally structured to minimize drastic changes to the existing .NET parsing code.

* feat: parse a large number of .NET metadata headers

Support parsing for most of the metadata headers described in ECMA 335 and update/add tests for them. This is intentionally structured to minimize drastic changes to the existing .NET parsing code.

* feat: parse a large number of .NET metadata headers

Support parsing for most of the metadata headers described in ECMA 335 and update/add tests for them. This is intentionally structured to minimize drastic changes to the existing .NET parsing code.